LayoutTests/media/media-controller-time-constant-expected.txt

This tests that currentTime is kept stable during script execution.
RUN(controller = video.controller)
EVENT(canplaythrough)
RUN(video.play())
RUN(controller.play())
EVENT(playing)
RUN(firstTime = controller.currentTime)
RUN(secondTime = controller.currentTime)
EXPECTED (firstTime - secondTime == '0') OK
END OF TEST
